html/semantics/embedded-content-0/the-video-element/video_018.htm

<!DOCTYPE HTML>
<html>
  <head>
    <title>HTML5 Media Elements: Seeking to a value less than 'initialTime' seeks to 'initialTime'.</title>
    <meta content="text/html; charset=UTF-8" http-equiv="Content-Type">
    <link rel="author" title="Microsoft" href="http://www.microsoft.com/" />
    <link rel="help" href="http://www.w3.org/TR/html5/embedded-content-0.html#dom-media-seek" />
    <meta name="assert" content="Seeking to a value less than 'initialTime' seeks to 'initialTime'." />
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <script type="text/javascript" src="../../../../common/media.js"></script>
    <script type="text/javascript">
      var test = async_test("Seeking to a value less than 'initialTime' seeks to 'initialTime'.", {timeout:5000});

      function do_play(event) {
        var vid = document.getElementById("video0");
        vid.pause();
        vid.currentTime = -1;
      }

      function do_seeked(event) {
        var vid = document.getElementById("video0");
        test.step(function() { 
          assert_equals(vid.initialTime, vid.currentTime);
          assert_equals(vid.initialTime, 0);
        } );
        test.done();
      }
    </script>
  </head>
  <body>
    <div id=log></div>
    <script type="text/javascript">

      document.write(
        "<video id='video0' src='" + getVideoURI("/media/movie_5") + "'" +
        " onseeked='do_seeked(event);'" +
        " onplay='do_play(event);'" +
        " autoplay></video>"
      );
    </script>
  </body>
</html>
